Create One_Time_Pad.py
The basic concept of the method  of One Time Pad is simple:
1) We take the text we want to encrypt and transform it into numbers
2) We choose a totally random key
3) We carry out a modular addition between the two

4) We communicate the message to the recipient

To decrypt the message, the recipient will only have to carry out the modular operation in reverse, using the same key.
The key, as you will have understood, must be as long as the text you want to encrypt. If you want to encrypt the text "HELLO MARIO" you will need a 10-character random key. To encrypt "NO" you need a 2-character key and so on.